CharacterConfigsRepositoryにqueryを移植
