+++ src/adapters/mod.rs
@@ -1,5 +1,7 @@
+mod character_configs_repository_impl;
 mod figure_records_repository_impl;
 mod user_configs_repository_impl;
 
+pub use character_configs_repository_impl::CharacterConfigsRepositoryImpl;
 pub use figure_records_repository_impl::FigureRecordsRepositoryImpl;
 pub use user_configs_repository_impl::UserConfigsRepositoryImpl;
