+++ src/commands/mod.rs
@@ -1,2 +1,1 @@
 pub mod character_config_seed_command;
-pub mod character_configs_command;
