+++ src/ports/mod.rs
@@ -1,7 +1,9 @@
+mod character_configs_repository;
 mod common;
 mod figure_records_repository;
 mod user_configs_repository;
 
+pub use character_configs_repository::*;
 pub use common::*;
-pub use figure_records_repository::FigureRecordsRepository;
-pub use user_configs_repository::UserConfigsRepository;
+pub use figure_records_repository::*;
+pub use user_configs_repository::*;
